requirements.txt
flask==3.1.1
python-dotenv==1.1.0
pypdf==5.6.1
sentence-transformers==4.1.0
langchain==0.3.25
langchain-core==0.3.77
langchain-community==0.3.25
langchain-text-splitters==0.3.8
langchain-openai==0.2.14
langchain-pinecone==0.2.0
langchain-huggingface==0.1.2
langchain-google-genai==2.0.0 
-e .